requirements.txt
arc-agi==0.9.6
google-genai>=1.72.0
python-dotenv>=1.1.0
anthropic==0.94.0
docker==7.1.0